Profiles are calculated correctly, and are now refactored
